refactor(services): eliminate pr → issue deep imports (Phase 6c)
Extract PR-specific cache operations from IssueTitleCacheService into
new FlowContextCacheService in shared/, and move resolve_issue_branch_input
from issue/branch_resolver to shared/branch_resolver.

Changes:
- NEW: src/vibe3/services/shared/context_cache.py (FlowContextCacheService)
- NEW: src/vibe3/services/shared/branch_resolver.py (moved from issue)
- MODIFY: src/vibe3/services/issue/branch_resolver.py (re-export only)
- MODIFY: src/vibe3/services/issue/title_cache.py (delegation to shared)
- MODIFY: src/vibe3/services/pr/service.py (use FlowContextCacheService)
- MODIFY: src/vibe3/services/pr/resolver.py (import from shared)
- MODIFY: src/vibe3/services/shared/__init__.py (register new exports)

This eliminates all 3 deep imports from pr → issue:
1. pr/service.py:133 → issue.title_cache.IssueTitleCacheService
2. pr/service.py:593 → issue.title_cache.IssueTitleCacheService
3. pr/resolver.py:10 → issue.branch_resolver.resolve_issue_branch_input

Backward compatibility:
- issue/branch_resolver.py preserves all exports via re-exports
- IssueTitleCacheService.update_pr() delegates to FlowContextCacheService
- check/pr_service.py continues to work via delegation

Tests: 74 passed, 1 xfailed, 1 xpassed

Closes #2577
