+++ src/vibe3/services/pr/resolver.py
@@ -8,7 +8,7 @@ from vibe3.clients import GitHubClient
 from vibe3.config import get_convention
 from vibe3.exceptions import UserError
 from vibe3.services.flow.service import FlowService
-from vibe3.services.issue.branch_resolver import resolve_issue_branch_input
+from vibe3.services.shared.branch_resolver import resolve_issue_branch_input
 
 
 def resolve_branch_from_pr(
